Cleanup & fmt api
