+++ api/src/xfr/asset_tracer.rs
@@ -1,16 +1,19 @@
-use crate::api::anon_creds::{Attr, AttributeCiphertext};
-use crate::xfr::structs::{AssetTracerDecKeys, AssetTracerEncKeys, TracerMemo};
-use crate::xfr::structs::{AssetType, ASSET_TYPE_LENGTH};
-use zei_algebra::bls12_381::{BLSScalar, BLSG1};
-use zei_algebra::prelude::*;
-use zei_algebra::ristretto::{RistrettoPoint, RistrettoScalar};
-use zei_crypto::basics::commitments::ristretto_pedersen::RistrettoPedersenGens;
-use zei_crypto::basics::elgamal::{
-    elgamal_decrypt, elgamal_decrypt_elem, elgamal_encrypt, ElGamalCiphertext, ElGamalDecKey,
-    ElGamalEncKey,
+use crate::anon_creds::{Attr, AttributeCiphertext};
+use crate::xfr::structs::{
+    AssetTracerDecKeys, AssetTracerEncKeys, AssetType, TracerMemo, ASSET_TYPE_LENGTH,
 };
-use zei_crypto::basics::hybrid_encryption::{
-    hybrid_decrypt_with_x25519_secret_key, hybrid_encrypt_with_x25519_key,
+use zei_algebra::{
+    bls12_381::{BLSScalar, BLSG1},
+    prelude::*,
+    ristretto::{RistrettoPoint, RistrettoScalar},
+};
+use zei_crypto::basics::{
+    commitments::ristretto_pedersen::RistrettoPedersenGens,
+    elgamal::{
+        elgamal_decrypt, elgamal_decrypt_elem, elgamal_encrypt, ElGamalCiphertext, ElGamalDecKey,
+        ElGamalEncKey,
+    },
+    hybrid_encryption::{hybrid_decrypt_with_x25519_secret_key, hybrid_encrypt_with_x25519_key},
 };
 
 pub type RecordDataEncKey = ElGamalEncKey<RistrettoPoint>;
